refactor: move ESG common xblock handlers to mixin
